fix: avoid mktemp
Remove patterns of creating a tmpfile (typically in `/tmp`) to only move
it when the operation complete. Instead, the operation is done directly
on the target file, with a possible fallback to delete the file if a
command fails.

The rationale is that `mktemp` creates files in `${TMPDIR:-/tmp}`. On
most systems, `/tmp` is a `tmpfs` filesystem with relatively limited
space available (especially on lower powered CI machines). This can be
an issue if dealing with large artefacts (for example, Rust build artefacts).

Signed-off-by: JP-Ellis <josh@jpellis.me>